requirements.txt
pandas
pymongo
python-dotenv
numpy
sqlalchemy
psycopg2-binary
pyspark==3.5.6
matplotlib
seaborn